fix: IntAbilityProperty에 대한 버프의 경우 실수 계산이 안되던 현상 수정
IntAbilityEffect에 대해서 Value의 자료형을 float으로 수정

+++ Status/Buff.cs
@@ -8,9 +8,9 @@ namespace MoraeGames.Library.Status
     public class IntAbilityEffect
     {
         [field: SerializeField] public int Id { get; private set; }
-        [field: SerializeField] public int Value { get; private set; }
+        [field: SerializeField] public float Value { get; private set; }
 
-        public IntAbilityEffect(int id, int value)
+        public IntAbilityEffect(int id, float value)
         {
             Id = id;
             Value = value;
